fix git

+++ database/migrations/2026_03_18_115712_drop_studies_columns.php
@@ -12,9 +12,16 @@ return new class extends Migration
     public function up(): void
     {
         Schema::connection(NON_BANKING_SERVICE_CONNECTION_NAME)->table('studies', function (Blueprint $table) {
-            $table->dropColumn('microfinance_loan_officer_count');
-            $table->dropColumn('consumerfinance_branches_count');
-            $table->dropColumn('consumerfinance_loan_officer_count');
+			if (Schema::connection(NON_BANKING_SERVICE_CONNECTION_NAME)->hasColumn('studies', 'microfinance_loan_officer_count')) {
+				$table->dropColumn('microfinance_loan_officer_count');
+			}
+			
+			if (Schema::connection(NON_BANKING_SERVICE_CONNECTION_NAME)->hasColumn('studies', 'consumerfinance_branches_count')) {
+				$table->dropColumn('consumerfinance_branches_count');
+			}
+			if (Schema::connection(NON_BANKING_SERVICE_CONNECTION_NAME)->hasColumn('studies', 'consumerfinance_loan_officer_count')) {
+				$table->dropColumn('consumerfinance_loan_officer_count');
+			}
         });
     }
